Add files via upload
Taleau COVID-19 India Dashboard Project" is a real-time data visualization platform that provides comprehensive insights into the COVID-19 situation in India. It offers a user-friendly interface with interactive charts, maps, and graphs to display the latest statistics related to COVID-19 cases, deaths, recoveries, and vaccination progress. The dashboard aims to help to make informed decisions in the fight against corona in India.